Add chat room ui & seperate UI from controller
